requirements.txt
aiogram==3.15.0
anthropic==0.43.0
aiosqlite==0.20.0
sqlalchemy[asyncio]==2.0.36
python-dotenv==1.0.1
openai==1.58.0
matplotlib==3.9.3
reportlab==4.2.5
PyPDF2==3.0.1
fastapi==0.115.0
uvicorn==0.32.0
PyJWT==2.10.1
